出品者か否かによる場合分け

+++ app/controllers/items_controller.rb
@@ -36,8 +36,12 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    unless current_user[:id] == item.user[:id]
+      redirect_to root_path
+    else
+      item.destroy
+      redirect_to root_path
+    end
   end
 
   private
